requirements/base.txt
beautifulsoup4==4.10
kink==0.6.2
python-dotenv==0.19.1
pytest==6.2.5
coverage==6.0.2
requests==2.26.0
python-slugify==5.0.2